+++ lib/features/shared/shared.dart
@@ -1,0 +1,1 @@
+export 'presentation/presentation.dart';
